Fix aliases not running any commands
Currently, `dest` in a sub-parser set is assigned to the sub-parser
that gets called. This also applies to aliases. When I used the base
sub-parser as the match condition, aliases just don't get matched as a
result. Now, I've abandoned the `dest` approach in favor of the
(officially) recommended solution: `set_defaults()` for setting the
function. However, as a consequence of the varying function signatures,
I have to add wrapper functions that use the arguments.